Refactored API handler organization
no refs

Refactored API handler organization to be more modular and easier to maintain:

- Grouped API handlers together by area of concern
- Moved API handlers to `src/http/api` directory
- Renamed API handlers to be more consistent with their area of concern
- Updated `src/mq/gcloud-pubsub-push/mq.ts` API handler naming
- Added docs for handlers
