--- src/api/index.ts
@@ -1,9 +1,0 @@
-export { getActivitiesAction } from './action/get-activities';
-export { getActivityThreadAction } from './action/get-activity-thread';
-export { profileGetAction } from './action/profile/get';
-export { profileGetFollowersAction } from './action/profile/get-followers';
-export { profileGetFollowingAction } from './action/profile/get-following';
-export { profileGetPostsAction } from './action/profile/get-posts';
-export { searchAction } from './action/search';
-export { webookPostPublishedAction } from './action/webhook/post-published';
-export { webhookSiteChangedAction } from './action/webhook/site-changed';
